% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explanatoryVarSampVsPop.R
\name{explanatoryVarSampVsPop}
\alias{explanatoryVarSampVsPop}
\title{Compare distribution of explanatoryVar in sample vs. population}
\usage{
explanatoryVarSampVsPop(samp, explanatoryVar)
}
\arguments{
\item{samp}{Output of simulateSample.}

\item{explanatoryVar}{The explanatory variable used to simulate occupancy and the sample.}
}
\value{
A ggplot object showing the density plots.
}
\description{
This function takes a sample inclusion matrix and an explanatory variable vector
and returns a ggplot object that shows the density distribution of the explanatory variable
for the population and each sample.
}
\examples{
samp <- list(sample_inclusion = matrix(sample(0:1, 100, replace = TRUE), nrow = 10))
explanatoryVar <- rnorm(100)
plot <- explanatoryVarSampVsPop(samp, explanatoryVar)
print(plot)
}
