% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulateDetection.R
\name{simulateDetection}
\alias{simulateDetection}
\title{Simulate Detection Histories Conditional on Occupancy and Sampling}
\usage{
simulateDetection(Y, S, w, alpha_p, beta_p, plot = TRUE)
}
\arguments{
\item{Y}{A matrix of true occupancy states for each site and year, dimensions \code{[nSites, nYears]}.}

\item{S}{A 3D array of visit indicators, dimensions \code{[nSites, nYears, nVisits]}.}

\item{w}{Numeric vector of detection covariates per site, length equal to \code{nSites}.}

\item{alpha_p}{Numeric. Intercept on the logit scale for detection probability.}

\item{beta_p}{Numeric. Slope for the detection covariate.}

\item{plot}{Logical. If TRUE, plots mean sampled-but-missed error and correlation of error with sampling (at least one visit=1; no visit =0).}
}
\value{
A list:
\itemize{
\item \code{D}: 3D array of detections \link{nSites x nYears x nVisits}
\item \code{error_summary}: data frame with \code{Year}, \code{mean_error}, \code{mean_sampled_missed}, \code{cor_sampling_error}
}
}
\description{
Simulate Detection Histories Conditional on Occupancy and Sampling
}
