% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulateOccupancy.R
\name{simulateOccupancy}
\alias{simulateOccupancy}
\title{Simulate Species Occupancy Over Time}
\usage{
simulateOccupancy(
  nYears,
  nSites,
  explanatoryVar,
  discreteExplanatoryVar = FALSE,
  beta0_extinction = NULL,
  beta0_colonization = NULL,
  beta0_initial = NULL,
  beta_extinction = NULL,
  beta_colonization = NULL,
  beta_initial = NULL,
  randomInitial = FALSE,
  randomInitialPsi = NULL,
  nBins = NULL,
  plot = "occupancy",
  colonisation_probs = NULL,
  extinction_probs = NULL,
  initial_occupancy_probs = NULL
)
}
\arguments{
\item{nYears}{Integer. The number of years over which to simulate occupancy.}

\item{nSites}{Integer. The number of sites to include in the simulation.}

\item{explanatoryVar}{Numeric vector. A vector of explanatory variables influencing occupancy, which must have length nSite.}

\item{discreteExplanatoryVar}{Logical. Whether the explanatory variable is categorical. The number of bins is set to the number of levels of the variable for categorical variables. This overrides the argument nBins.}

\item{beta0_extinction}{Numeric. Intercept for the extinction probability.}

\item{beta0_colonization}{Numeric. Intercept for the colonization probability.}

\item{beta0_initial}{Numeric. Intercept for the initial occupancy probability.}

\item{beta_extinction}{Numeric. Slope for the extinction probability.}

\item{beta_colonization}{Numeric. Slope for the colonization probability.}

\item{beta_initial}{Numeric. Slope for the initial occupancy probability.}

\item{randomInitial}{Logical. Whether to generate initial occupancy randomly. If false, then beta parameters must be set.}

\item{randomInitialPsi}{Numeric. If \code{randomInitial} is \code{TRUE}, the probability of initial occupancy.}

\item{nBins}{Integer. The number of bins to divide the explanatory variable into.}

\item{plot}{Character. Specifies the plot type to generate: "occupancy" for overall occupancy or "bin" for occupancy by bin.}

\item{extinction_probs}{Numeric. A vector of extinction probabilities. Must be of length nSites. If supplied, extinction probabilities will not be modelled as a function of explanatoryVar.}

\item{initial_occupancy_probs}{Numeric. A vector of initial occupancy probabilities for the first year. Must be of length nSites. If supplied, initial occupancy probabilities will not be modelled as a function of explanatoryVar.}

\item{colonisaton_probs}{Numeric. A vector of colonisation probabilities. Must be of length nSites. If supplied, colonisation probabilities will not be modelled as a function of explanatoryVar.}
}
\value{
A list containing:
\itemize{
\item \code{occupancy_all_sites}: A data frame of overall occupancy probabilities over time.
\item \code{occupancy_per_bin}: A data frame of bin-wise occupancy probabilities over time.
\item \code{simulated_data}: A matrix of simulated occupancy states for each site and year.
}
}
\description{
This function simulates species occupancy over time for a given number of sites and years,
incorporating explanatory variables and probabilities for extinction and colonization.
}
\examples{
sim_data <- simulateOccupancy(
 nYears = 10,
 nSites = 500,
 explanatoryVar = runif(500, min = 0, max = 1),
 beta0_extinction = -1,
 beta0_colonization = -3,
 beta0_initial = NULL,
 beta_extinction = -2,
 beta_colonization = 2,
 beta_initial = NULL,
 randomInitial = TRUE,
 randomInitialPsi = 0.3,
 nBins = 5,
 plot = "occupancy"
)

true_occupancy <- sim_data$occupancy_all_sites
occupancy_by_bin <- sim_data$occupancy_per_bin

# Further analysis or visualization can follow

}
